requirements.txt
crewai
crewai-tools
python-dotenv
langchain-groq
langchain-openai
#serper-api-key
streamlit
langchain-community
pydantic